crud operations on jobs

+++ src/main/java/com/example/greencommute/controller/AdminConstructor.java
@@ -1,8 +1,10 @@
-package com.example.greencommute.rest;
+package com.example.greencommute.controller;
+
 
 import com.example.greencommute.entity.Job;
 import com.example.greencommute.entity.Skill;
 import com.example.greencommute.entity.User;
+import com.example.greencommute.exception.JobNotFoundException;
 import com.example.greencommute.service.JobService;
 import com.example.greencommute.service.SkillService;
 import com.example.greencommute.service.UserService;
@@ -12,38 +14,35 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.logging.Logger;
 
-@RestController
+@RequestMapping("/admin")
 @Slf4j
-public class GreenCommuteController {
-
+@RestController
+public class AdminConstructor {
     @Autowired
     private UserService userService;
-
     @Autowired
     private SkillService skillService;
 
     @Autowired
     private JobService jobService;
 
-    @GetMapping("/users")
-    public List<User> findUsers(){
-        return userService.findAll();
-    }
-
     @GetMapping("/jobs")
     public List<Job> findAllJobs(){
 
-        log.info(" got all the jobs");
         return jobService.findAllJobs();
     }
 
     @GetMapping("/jobs/{jobId}")
     public Optional<Job> findJobById(@PathVariable int jobId){
-        return jobService.findJobById(jobId);
-    }
 
+        Optional<Job> theJob=jobService.findJobById(jobId);
+
+        if(theJob==null){
+            throw new JobNotFoundException("Given Id does not exist");
+        }
+        return theJob;
+    }
 
 
     @PostMapping("/jobs")
@@ -52,8 +51,9 @@ public class GreenCommuteController {
         List<Skill> skills=theJob.getSkills();
 
         if(skills!=null){
-            for(Skill skill:skills)
+            for (Skill skill:skills) {
                 skillService.saveSkill(skill);
+            }
         }
         log.info(" job id :"+theJob.getJobId()+"job name :"+theJob.getJobName()+"the skills"+theJob.getSkills());
 
@@ -67,13 +67,28 @@ public class GreenCommuteController {
 
     @DeleteMapping("/jobs/{jobId}")
     public void deleteJobById(@PathVariable int jobId){
+
+        Optional<Job> theJob=jobService.findJobById(jobId);
+
+        if(theJob==null){
+            throw new JobNotFoundException("Given Id does not exist");
+        }
         jobService.deleteJob(jobId);
     }
 
-    @GetMapping("/jobs/location={location}")
-    public List<Job> getJobsByLocation(@PathVariable("location") String location){
-
+    @GetMapping("/jobs/location")
+    public List<Job> getJobsByLocation(@RequestParam("location") String location){
         return jobService.getJobsByLocation(location);
+    }
+
+    @GetMapping("/jobs/skills")
+    public List<Job> getJobsBySkills(@RequestParam("skill") String skill){
+        return jobService.getJobsBySkills(skill);
+    }
+
+    @GetMapping("/users")
+    public List<User> findUsers(){
+        return userService.findAll();
     }
 
 }
